heritage chatbot upload/backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
pydantic
pymongo
bcrypt
email-validator

certifi
dnspython
